requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
pyTelegramBotAPI==4.11.0
python-dotenv==0.21.1
python-telegram-bot==13.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.30.0
six==1.16.0
tornado==6.2
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
